# docs/audit-archive/openapi.yaml
openapi: 3.0.3
info:
  title: UmbraPL API
  description: API для платформы управления пользователями и процессорами UmbraPL
  version: 1.0.0
  contact:
    name: UmbraPL Team
    email: admin@umbra-platform.dev
  license:
    name: Private
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://umbra-platform.onrender.com/api
    description: Production server
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: Вход в систему
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Успешная аутентификация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: auth-token
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: admin@umbra-platform.dev
        password:
          type: string
          minLength: 6
          example: umbra2024
...
    LoginResponse:
      type: object
      properties:
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        isBlocked:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserRole:
      type: string
      enum:
        - USER
        - ADMIN
        - MODERATOR
        - MEDIA_BUYER
        - SUPPORT
        - PROCESSOR
        - ROP_PROCESSOR
        - ROP_BUYER
        - BUYER
        - LEAD_BUYER
        - FINANCE
    UserStatus:
      type: string
      enum:
        - PENDING
        - APPROVED
        - REJECTED
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        errors:
          type: object
          additionalProperties: true
  responses:
...
    Unauthorized:
      description: Требуется авторизация
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Недостаточно прав
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...